Axis class now has a separate value and raw value
A (maybe) important change: now the axis detector that takes precedence
is the last one that has a *non-zero* value (not accounting for the
deadzone). Because of this, it is recommended that binary axis detectors
take precedence, as analog axis detectors will almost always be
registering an input due to almost never being perfectly centered.